Bump Dojo v1.2.1 (#21)
* Update dojo dependency to version 1.2.1 in Scarb.lock, Scarb.toml, and CI workflow configuration

* Add .tool-versions file and update Game struct id type from u32 to u64

* Update game_id type from u32 to u64 across multiple event and model structs

+++ src/events/card_play_event.cairo
@@ -18,7 +18,7 @@ enum EventType {
 pub struct CardPlayEvent {
     #[key]
     pub player: ContractAddress,
-    pub game_id: u32,
+    pub game_id: u64,
     pub mod_id: felt252,
     pub event_type: EventType,
     pub special: Span<(u32, i32)>,
